Sprint and Flashlight Alpha Improvements
Added screen shake to the player while sprinting. Fixed a bug that allowed the player to sprint while not moving. Increased the brightness and range of the flashlight. Revamped the battery system to have 3 battery types, small, medium, and large. Each size fills the battery meter in 1/3rds. Etc small battery is 1/3 fill and large battery is 3/3 fill. Changed temp UI to show rounded time remaining for both the flashlight and sprint meter.