Altered VO and added the ability to toggle subtitles
Subtitles can now be enabled or disabled from the mainmenu. VO will also stop when the game is paused. The subtitle system is very crude, as a one that would show each line individually would require the VO to be cut up into chunks or to manually time it out.